Refactor: extracted some methods #7
